+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "vite-starter-prendy",
-  "version": "0.1.4",
+  "version": "0.1.5",
   "homepage": ".",
   "repository": {
     "type": "git",
@@ -17,35 +17,34 @@
   "license": "MIT",
   "scripts": {
     "dev": "vite",
-    "build": "vite build",
-    "buildprev": "tsc && vite build",
+    "build": "tsc && vite build",
     "preview": "vite preview"
   },
   "dependencies": {
-    "@babylonjs/core": "6.5.0",
-    "@babylonjs/gui": "6.5.0",
-    "@babylonjs/inspector": "6.5.0",
-    "@babylonjs/loaders": "6.5.0",
-    "chootils": "^0.3.8",
+    "@babylonjs/core": "6.14.1",
+    "@babylonjs/gui": "6.14.1",
+    "@babylonjs/inspector": "6.14.1",
+    "@babylonjs/loaders": "6.14.1",
+    "chootils": "^0.3.9",
     "delay": "^5.0.0",
-    "repond": "^0.9.1",
-    "repond-movers": "^0.6.1",
-    "prendy": "^0.9.5",
+    "prendy": "^0.9.13",
     "react": "^18.2.0",
     "react-babylonjs": "3.1.17",
     "react-dom": "^18.2.0",
     "react-promise-suspense": "^0.3.3",
     "react-spring": "^9.6.1",
+    "repond": "^0.9.2",
+    "repond-movers": "^0.7.0",
     "stringz": "^2.1.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
     "@vitejs/plugin-react": "^2.0.0",
-    "typescript": "^4.7.4",
-    "vite": "^3.0.2",
+    "typescript": "^5.1.6",
+    "vite": "^4.3.9",
     "vite-plugin-pwa": "^0.14.4",
-    "vite-tsconfig-paths": "^3.5.0"
+    "vite-tsconfig-paths": "^4.2.0"
   },
   "prettier": {
     "printWidth": 120
