+++ src/assets/assets.ts
@@ -13,7 +13,7 @@ import {
   modelNames,
 } from "./models/models";
 import { MusicName, musicFiles, musicNames } from "./music/music";
-import { FontName, PRENDY_OPTIONS, fontNames } from "./options";
+import { PRENDY_OPTIONS } from "./options";
 import { PickupName, pickupsInfo } from "./pickups/pickups";
 import {
   AnyCameraName,
@@ -32,6 +32,7 @@ import {
 } from "./places/places";
 import { SoundName, soundFiles, soundNames } from "./sounds/sounds";
 import { SpeechVidName, speechVidFiles } from "./speechvids/speechvids";
+import { FontName, fontNames } from "./fonts";
 
 export const PRENDY_ASSETS = {
   placeInfoByName,
@@ -49,28 +50,15 @@ export const PRENDY_ASSETS = {
   musicFiles,
   soundFiles,
   speechVidFiles,
-  // startOptions
-};
-
-type ModelNamesByPlaceLoose = Record<PlaceName, ModelName[]>;
-
-type PlaceInfoByName = typeof placeInfoByName;
-type ModelInfoByName = typeof modelInfoByName;
-type DollOptions = typeof dollOptions;
-type CharacterOptions = typeof characterOptions;
-type MusicFiles = typeof musicFiles;
-type SpeechVidFiles = typeof speechVidFiles;
-type SoundFiles = typeof soundFiles;
-type PickupsInfo = typeof pickupsInfo;
-
-type PrendyOptions = typeof PRENDY_OPTIONS;
+  prendyOptions: PRENDY_OPTIONS,
+} as const;
 
 export type LocalAssetsTypes = {
-  PrendyOptions: PrendyOptions;
-  PlaceInfoByName: PlaceInfoByName;
-  ModelInfoByName: ModelInfoByName;
-  DollOptions: DollOptions;
-  CharacterOptions: CharacterOptions;
+  PrendyOptions: typeof PRENDY_OPTIONS;
+  PlaceInfoByName: typeof placeInfoByName;
+  ModelInfoByName: typeof modelInfoByName;
+  DollOptions: typeof dollOptions;
+  CharacterOptions: typeof characterOptions;
   ModelName: ModelName;
   DollName: DollName;
   CharacterName: CharacterName;
@@ -95,20 +83,18 @@ export type LocalAssetsTypes = {
   BoneNameByModel: BoneNameByModel;
   MaterialNameByModel: MaterialNameByModel;
   MeshNameByModel: MeshNameByModel;
-  ModelNamesByPlaceLoose: ModelNamesByPlaceLoose;
+  ModelNamesByPlaceLoose: Record<PlaceName, ModelName[]>;
   SegmentNameByPlace: SegmentNameByPlace;
-  MusicFiles: MusicFiles;
-  SpeechVidFiles: SpeechVidFiles;
-  SoundFiles: SoundFiles;
-  PickupsInfo: PickupsInfo;
+  MusicFiles: typeof musicFiles;
+  SpeechVidFiles: typeof speechVidFiles;
+  SoundFiles: typeof soundFiles;
+  PickupsInfo: typeof pickupsInfo;
 };
-
-type LocalStoreHelpers = typeof storeHelpers;
 type LocalStores = typeof stores;
+type LocalStoreHelpers = typeof storeHelpers;
 
 declare module "prendy" {
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
   interface CustomAssetsTypes extends LocalAssetsTypes {}
-  interface CustomStoreHelpers extends LocalStoreHelpers {}
   interface CustomPrendyStores extends LocalStores {}
+  interface CustomStoreHelpers extends LocalStoreHelpers {}
 }
