--- src/assets/places/basement/join_depth_vids.txt
@@ -1,4 +1,0 @@
-# tells the video joiner the names and order of the depth videos 
-file 'first_cam_start_depth.mp4'
-file 'waterfall_cam_start_depth.mp4'
-
