--- src/assets/places/stairy/join_color_vids.txt
@@ -1,6 +1,0 @@
-# tells the video joiner the names and order of the color videos 
-file 'room_camera_start.mp4'
-file 'left_door_camera_start.mp4'
-file 'downstair_camera_start.mp4'
-file 'stairs_camera_start.mp4'
-
