+++ src/main.tsx
@@ -1,5 +1,4 @@
 import { PRENDY_ASSETS } from "assets/assets";
-import { PRENDY_OPTIONS } from "assets/options";
 import { makeDebugFrameRate, makePrendyApp } from "prendy";
 import React from "react";
 import ReactDOM from "react-dom";
@@ -9,7 +8,6 @@ import "./index.css";
 
 const PrendyApp = makePrendyApp({
   prendyAssets: PRENDY_ASSETS,
-  prendyOptions: PRENDY_OPTIONS,
   stores,
   storeHelpers,
   customRules,
@@ -19,7 +17,7 @@ const DebugFrameRate = makeDebugFrameRate(storeHelpers);
 
 ReactDOM.render(
   <React.StrictMode>
-    <PrendyApp>{/* Add custom components here, like minigames or particles */}</PrendyApp>
+    <PrendyApp>{/* Add custom components here,like minigames or particles */}</PrendyApp>
     {/* <DebugFrameRate /> */}
   </React.StrictMode>,
   document.getElementById("root")
