+++ src/rules/places.ts
@@ -1,4 +1,5 @@
-import { prendy } from "stores/stores";
+import delay from "delay";
+import { prendy, showMiniBubble } from "stores/stores";
 
 export const placeLoadRules = prendy.rules.makePlaceLoadRules(
   // at start of any places
@@ -16,11 +17,14 @@ export const placeLoadRules = prendy.rules.makePlaceLoadRules(
   {
     async stairy({ storyState: { exampleStoryToggle } }) {
       prendy.story.players.takePickup("key");
+
+      await delay(500);
+      showMiniBubble("‼️");
     },
   }
 );
 
-export const placeNotLoadedRules = prendy.rules.makePlaceNotLoadedRules({
+export const placeUnloadRules = prendy.rules.makePlaceUnloadRules({
   async stairy() {
     // hideDoll("walker");
   },
